feat: DialogManager 디렉토리 변경 및 모달 애니메이션 추가
