+++ src/components/common/DialogManager/AddCandy/CandyAdded.tsx
@@ -2,8 +2,8 @@ import React, { useState, useMemo } from 'react';
 import Image from 'next/dist/client/image';
 import styled from 'styled-components';
 import { useAtom } from 'jotai';
-import { openCandyModal } from '../../../states';
-import Button from '../../common/Button';
+import { openCandyModal } from '../../../../states';
+import Button from '../../Button';
 import AddCandyDate from './AddCandyDate';
 
 const MainBox = styled.div`
