+++ src/components/common/DialogManager/AddCandy/index.stories.tsx
