+++ src/components/common/DialogManager/AddCategory/CategoryAdded.tsx
@@ -3,8 +3,8 @@ import Link from 'next/link';
 import styled from 'styled-components';
 import Image from 'next/dist/client/image';
 import { useAtom } from 'jotai';
-import { openCategoryModal } from '../../../states';
-import Button from '../../common/Button';
+import { openCategoryModal } from '../../../../states';
+import Button from '../../Button';
 
 const MainBox = styled.div`
   display: flex;
